Warn if non-relative, since cross-origin download has many problems (#4779)
This PR fix #4723 to avoid a common pitfall that is cross-origin
download on the browser-side.

Due to browser security reasons outlined
https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/a#download,
if it is cross-origin, then the browser ignores the `download` and thus
`filename` and `media_type` alike.

The conclusion from #4723 appears to be, there is no legitimate use for
running `ui.download.from_url` from a cross-origin URL. Since it'll only
work as the user expected only if:
- Server responds a file type the browser can't otherwise open a new
tab,
- User did not pass filename (which would otherwise be ignored). 

Thus, it is recommended that either: 
- User change their code to relative URL
- Or if they insist, simply use `ui.navigate.to` with `new_tab = True`
(https://nicegui.io/documentation/navigate#ui_navigate_to_(formerly_ui_open))

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>