UserInteraction.click(): Handle radio elements (#4849)
### Motivation

<!-- What problem does this PR solve? Which new feature or improvement
does it implement? -->
Currently, `UserInteraction.click()` does not support `ui.radio`
elements. This PR updates it so radio elements work similarly to
`ui.select` elements.

<!-- Please provide relevant links to corresponding issues and feature
requests. -->

### Implementation

<!-- What is the concept behind the implementation? How does it work?
-->
It adds a new section of logic to `UserInteraction.click()` for radio
elements.

<!-- Include any important technical decisions or trade-offs made -->

I have not added documentation for this case, but it's so similar to
`ui.select` that perhaps no additional documentation is needed. (But I'm
happy to add documentation if that would be useful.)

### Progress

- [X] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [X] The implementation is complete.
- [X] Pytests have been added (or are not necessary).
- [X] Documentation has been added (or is not necessary).

---------

Co-authored-by: Barry Hart <barry.hart@zoro.com>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>