Fix deselection of table rows by using the correct row key (#4893)
### Motivation

Discovered the bug in Issue #4875 where rows in a table can be selected,
but not deselected. This is due to the function `handle selection` which
refers to the row_key which is given in the initialisation of the table,
but if it is assigned afterwards, it does not update in the function.
Therefore replacing the row_key with self.row_key, to ensure that the
class variable is used.

Co-authored-by: tobias hugentobler <tobias.hugentobler@stadlerrail.com>