Make binding transform functions optional (#4910)
### Motivation

I noticed that the `bind_to` and `bind_from` functions in the binding
module require the user to specify transform functions - in contrast to
the two-way `bind` which defaults to identity functions `lambda x: x`.
(Note that the user usually uses special binding methods like
`bind_value` or `bind_text`, so this deficit remained unnoticed for a
while.)
Apart from that it would be more efficient to allow `transform=None` and
skip the function call in this case.

### Implementation

This PR makes all binding transform functions `Optional` and sets the
default `None`. The docstrings are updated to mention the default
behavior.
When using transform functions in binding.py, the possible is value
`None` is handled with an extra if-condition.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.