Fix nested propagation of bindable properties (#4933)
### Motivation

We noticed that changing the value of a bindable property multiple times
within the change handler of another bindable property doesn't work
correctly:

```py
class Demo:
    a = binding.BindableProperty()
    b = binding.BindableProperty(on_change=lambda obj, _: obj.change_a())

    def __init__(self) -> None:
        self.a = 0
        self.b = 0

    def change_a(self) -> None:
        self.a = 1
        self.a = 2

demo = Demo()
ui.label().bind_text_from(demo, 'a', lambda a: f'a = {a}')  # stuck at 1
ui.number().bind_value_to(demo, 'b')  # should set a to 1 and then 2
```

### Implementation

We correctly assumed that the context variable `propagation_visited`
wasn't used correctly. At the first assignment to `self.a` the property
is added to the set of visited properties. Then the propagation of the
second assignment falsely assumes that `self.a` has already been
propagated.

Now we make sure to always open a new propagation context whenever a
value assignment or a refresh step of the binding loop triggers a
propagation. This eliminates the need for distinguishing between root
and non-root calls.

This PR also adds a pytest to cover this case.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation is not necessary.