Let `ui.header` update `scrollPaddingTop` more nicely with a custom header.js component (#4937)
### Motivation

In
https://github.com/zauberzeug/nicegui/issues/4912#issuecomment-3035026068
we noticed that `ui.header` is adding code blocks to the HTML body which
are not cleaned up when recreating the header. This is concerning in
single-page applications where one might want to countlessly recreate
the whole UI including page layout elements.

### Implementation

1. This PR splits the old page_layout.py into separate files for the
individual elements.
2. A custom header.js component takes care of creating the resize
observer if needed.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.