New Element - InputChips (#4940)
### Motivation

Note that #4927 can not be resolved since Quasar framework has same
behavior as well. The root cause lies in the dropdown itself which is
directly coupled with the behavior of `options`. Setting props
`hide-dropdown-icon popup-content-class='hidden' use-chips` does not
help either. You may accidentally toggle the first element if a
substring is typed in (#4927 shows some clips on this behavior). This
will be troublesome in long lists.

`ui.select` is at the moment can be an alias for `Dropdown Selection`.
Since options is more or less the backbone of ui.selecct, making the
dropdown optional may break ui.select.

Therefore, I created a new element `InputChips` to overcome the issue
for users who want `new_value_mode` and safely use it. Using this
element will add values to list with input and removing can be done with
chip's clear button rather than dropdown. Having an element like this
ready to use would be helpful for myself (because this was exactly what
I wanted to create a form) or users like #1649.

### Implementation

- It replaces `ChoiceElement` with `ValueElement` to remove the
dependency on options/dropdown and the JS file contains barebones of
Qselect.
- It converts the value label dict from Qselect to list of values and
compares with the new values. So this simplifies backend work done to
keep the values validated according to `new_value_mode` set for the
element.
- Unlike my previous PR, `use_delimiter` works for all cases since this
only deals with list.

Partly I feel there will be better performance for larger list since
this element quickly deals with the list of values in transit while
ui.select have to run `_hande_new_value` after. Remember that
`_hande_new_value` runs multiple times if `use_delimiter` is True, for
cases like 2+ values are added in one go. (Though this will be the case
if #4936 or this PR is merged at some point to use `use_delimiter`)

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation has been added.

I created this draft PR before starting pytest/docs, to get some
opinions/thoughts of this feature whether it will be good addition or
not.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>