Add rate limits for hosted website (#4987)
### Motivation

https://nicegui.io sometimes sees to much traffic from single users.
This can lead to bad experience for everyone else.

### Implementation

This pull request sets rate limits via
https://github.com/laurentS/slowapi. I deliberately just made it an
optional feature and added the `slowapi` package to our fly.io
deployment. If it works we might make it a general feature in NiceGUI.
But thats for an eperate PR.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have are not necessary.
- [x] Documentation is not necessary.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>