Improve documentation page for `ui.sub_pages` (#5000)
### Motivation

We noticed that embedded `ui.sub_pages` demos caused the whole
documentation page to scroll to the top, which made exploring the demos
rather tedious and confusing. Apart from that, navigating to /other/a in
one demo caused 404 errors in other demos with no inherent way to get
the demo working again.

### Implementation

We decided to stop using real `ui.sub_pages` elements for demos but to
mock them with a `FakeSubPages` element. The implementation is pretty
straightforward: Initialize it with routes and optional data, create a
`link` with a given text and route and call `init` at the place where
the subpages should be rendered. A `ui.timer` is used to call both sync
and async subpage functions.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.