Fix routing to other FastAPI pages in documentation (#5004)
### Motivation

As @evnchn noted in #5001, migrating the documentation to a single-page
application (#4902) introduced a bug that makes the FastAPI pages
unreachable. The issue was caused by the FastAPI catch-all route for the
SPA entry point.

### Implementation

As an temporary fix, this PR uses explicit routes for the SPA plus a
custom `ui.sub_pages` implementation which short-stops rendering the
main page from rendering before the SPA can decide to route to another
FastAPI route.
A better solution might be possible if we implement #4964 in 3.0.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have are not necessary.
- [x] Documentation is not necessary