Introduce `ValueChangeEventArguments.previous_value` (#4456)
Inspired by discussion #4410, this PR introduces a `previous_value`
field for `ValueChangeEventArguments`. This change is pretty
straightforward and only requires adjustments for two mixins and two
elements.

Right now it is lacking backward-compatibility though. If user code
derives custom `ValueChangeEventArguments`, they are now missing the
`previous_value` field. Should we add a default value? But it should be
clear that this default is _not_ the previous value, but "not set". We
might need a sentinel...

And in 3.0 I'd like to enforce the `previous_value` field. A deprecation
warning for the default value would be great, but probably hard to
achieve.