Use fly.io autoscaler for https://nicegui.io (#5027)
### Motivation

Managing fly.io instances has been quite a burden. Sometimes it was
simpler to keep more instances running than necessary to handle spikes
at other times. The fly.io build in auto-scaler could not be used out of
the box, because the cold-start of a machine in a region without any
instances takes a few seconds (which results in waiting time for the
visitor).

### Implementation

The new `deploy.py` keeps one instance per region running by removing it
from the autoscaler. With that trick, instances are started and stopped
by fly.io on demand without having poor experience for regions with only
a few visitors.


### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation has been added.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>