Allow canceling the currently executed invoctation of ui.timer (#5011)
### Motivation

Originally, this PR was motivated by #5010, where the async demo in
`ui.sub_pages` documentation does not work properly when rapidly
clicking the links. When trying to fix the bug, I discovered that there
is currently not an option in `ui.timer` to cancel the ongoing task.

### Implementation

This PR introduces `timer.cancel(with_current_invocation=True)` to allow
canceling the internal awaited coroutine.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation has been added.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>