Use weak references for parent slot to avoid cyclic references (#5053)
### Motivation

In #5052 we noticed that there are still cyclic references between
elements and their parent slots. To reduce work for the garbage
collector and use much cheaper reference counting, we should avoid such
cycles if possible.

### Implementation

Similar to other weak references, this PR introduces a private attribute
`self._parent_slot` and a public property `self.parent_slot`. Internal
access to the parent slot is optimized to reduce the number of
dereferentiations.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary): I'm not sure how
because we'd need to disconnected the client and wait.
- [x] Documentation has been added (or is not necessary).