Close unused user storages (#5074)
### Motivation

As was discovered in
https://github.com/zauberzeug/nicegui/issues/4998#issuecomment-3217815184,
each request creates a user storage object -- which is never removed.
This results in a gradual memory increase.

### Implementation

I created `prune_user_storage` which collects all active user storage
sessions from current list of clients and then remove all user storages
which are not in that list. Because of a cyclic dependency to client, I
placed the prune method not in storage.py but in nicegui.py. For
consistency I also moved the `prune_tab_storage` method.

I also removed the code from #4982 which closed non-page requests. But
that code forgot to remove the storage object and only closed it. And
removing it is not ok, because another client might still require it for
access.

To improve testability and less code overall I also changed the pruning
methods from `while True` to `app.timer`.

And I discovered that `client.tab_id` was not set to `None` on websocket
disconnect. This is bad because it lets `client.has_socket_connection`
wrongly return `True`.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation is not necessary

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>