Make props, classes and style observable to send updates automatically (#4978)
### Motivation

When reviewing PR #4969, we noticed that it would be handy if the props
dictionary would automatically update the frontend when changed:
```py
b = ui.button('test', on_click=lambda: b.props.update(icon='face'))
```

### Implementation

This PR replaces the `dict` base class with an `ObservableDict` and
calls a new `_update` method. In order to avoid keeping strong reference
to `element`, this method needs to use the weak reference
`self._element`.

I'm marking the PR as draft because we should carefully evaluate memory
and performance, add pytests and documentation. Although the change
shouldn't break anything, it might be a good idea to release this
feature in 3.0.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Introduce `@helpers.prevent_recursion`
- [x] Existing pytests are passing.
- [x] Pytests have been added.
- [x] Documentation has been added.
- [x] This PR breaks
https://nicegui.io/documentation/table#show_and_hide_columns [BREAKING
CHANGE].
- [x] The condition `if self._send_update_on_value_change:` might be
ineffective now.
- [x] `test_apply_format_on_blur` is red