Fix direct opening of async nested sub pages (#5089)
### Motivation

As @alexzaech reported in #5085, opening a nested sub page which is
inside a async sub page shows the 404 error page.

### Implementation

Currently this PR only reproduces the problem in a screen test. 
A minimal reproduction:

```py
from nicegui import ui

@ui.page('/')
@ui.page('/{_:path}')
def index():
    ui.sub_pages({'/': main})

async def main():
    ui.sub_pages({
        '/': sub_main,
        '/sub': sub_sub_page,
    })

def sub_main():
    ui.link('Go to sub page', '/sub')

def sub_sub_page():
    ui.label('sub page')

ui.run()
```

If you access `/` everything works fine. And clicking on "Go to sub
page" also works. But if you directly open `/sub` you see a 404 page.

The problem is this: due to the `async` keyword the execution of `main`
is a coroutine which is not executed until the event loop ticks it
forward. Hence the sub pages element in main is not available and


https://github.com/zauberzeug/nicegui/blob/78bf3151b645573834e46aecd8e2c50977d4ac26/nicegui/elements/sub_pages.py#L111-L115

determines that the 404 should be shown. It works for the `/` because
`match.remaining_path` is empty.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] try to find a way to allow 404 http errors **and** evaluate async
builder functions properly
- [ ] verify that the approach will still work with 3.0 (especially
#5005, where SPA root page is processed in 404 handler)
- [x] Pytests have been added.
- [x] Documentation is not necessary.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>