add an option to bindings to check if target exists (#5040)
### Motivation

When refactoring or renaming variables their string names in bindings
are often missed since the rename IDE action doesn't pick it up. We
discussed creating a method that gets the variable name as a string from
a reference which proved possible but costly. As an alternative to raise
awareness to missed variable names after renaming we propose to add an
option that checks if the binding target exists and warns if it's not.

### Implementation

The implementation is rather straight forward and involves a lot of
trivial parameter adding. The check takes place in the functions `bind`,
`bind_to` and `bind_from` in `bindings.py` which are used in all the
other occurrences.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
  - [x] update all docstrings
  - [x] `direction` could be a `Literal`
  - [x] check both, source and target
- [x] change it to `check: bool | None = None` and check non-dicts by
default
- [x] Pytests have been added.
- [x] Documentation has been added.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>