Close sub_pages render coroutine on cancellation (#5145)
### Motivation

As @platinops described in #5142, using an async function as `/` and
then calling `sub_pages.add` results in a "coroutine was never awaited"
warning. Reproduction:

```py
@ui.page('/')
@ui.page('/{_:path}')
def index():
    pages = ui.sub_pages({'/': main})
    pages.add('/other', lambda: ui.label('other page'))

async def main():
    ui.label('main page')
```

### Implementation

If the renderer task was cancelled before it could start, we must close
it. I also added a test, but discovered that it's just a warning. After
some experiments I decidet to not added "warnings as errors" to the
`pyproject.toml`. So the test passes but prints a warning. We can turn
on "warning as errors" in a different PR (targeting 3.0).

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation is not necessary.