Call lifecycle event handlers more consistently (#5173)
### Motivation

In #4285 we noticed that connect and disconnect handlers are not both
called when reconnecting. To avoid breaking changes, we left it
unchanged until now. With version 3.0 we now have the chance to unify
how lifecycle event handlers are called.

### Implementation

This PR introduces new event handlers to complete the following list:

- `on_handshake`: when the first socket connection is established
- `on_connect`: when the socket connection is established (even during
reconnect)
- `on_disconnect`: when the socket connection is interrupted (even
during reconnect)
- `on_deleted`: when a client is deleted after it didn't reconnect for a
while

Note that the (internal) `client.handle_handshake` method is called on
every reconnect, but the handshake even handler is only called at the
first connection.

I also added a new pytest, but didn't manage to simulate connection
losses.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (although without maximum test coverage).
- [x] Documentation has been added.
- [ ] decision about event names, especially "on_handshake"