Persisting on teardown fails on 3.9 (#5182)
### Motivation

Working on #5178 uncovered the brittle implementation of the
`background_tasks.await_on_shutdown` annotation done in #4641.
Especially on Python 3.9.

### Implementation

This PR suggests to use a simple list of "task to be awaited" instead of
the weak ref set. In addition the detection of the decorator is now
implemented in a more robust and standard way via class wrapping.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation is not necessary.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>