Treat pytest warnings as errors (#5186)
### Motivation

This PR is based on @evnchn's PR #5153 which activates "warnings as
errors" for pytests.

### Implementation

To fix some of the new errors, we had to fix the teardown of background
tasks (PR #5182) and introduce a new API for `ui.upload` (PR #5178).
This also led to a rework of some of our examples related to
`ui.upload`.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation is not necessary.

---------

Co-authored-by: evnchn <evanchan040511@gmail.com>
Co-authored-by: Rodja Trappe <rodja@zauberzeug.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>