Re-initialize `process_pool` if it was broken by a bad `run.cpu_bound` (#5129)
### Motivation

Fixes #4980, where we see if the process pool is broken in a manner of
process termination, there is no way for the user to recover it without
resorting to directly writing to `run.process_pool =
ProcessPoolExecutor()`

### Implementation

This implementation checks if the process pool is broken after an error
has been raise by asking it to run an empty lambda. If that fails then
the process pool is broken and we re-initialize it. Nevertheless the
error will still be raise.

Effect is, no matter how mangled your calls to `run.cpu_bound`, you can
never wreck it and affect the next normal call.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
 - Not sure if I should add some. 
- [x] Documentation is not necessary.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>