Introduce new `sanitize` argument for `ui.html` (#5187)
### Motivation

It has been pointed out to us that `ui.html` could be considered unsafe
because it doesn't sanitize its content. It could be filled with user
input, opening the door for nasty things.

### Implementation

Therefore we decided to introduce a new required `sanatize` parameter
which either accepts a callable sanitization function like
`Sanitizer().sanitize` from the `html-sanitizer` package, or `False` to
keep the content unchanged. This way we force the developers to think
about the intention of `ui.html` elements.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
    - [x] What to do with `ui.chat_message`?
- [x] Pytests have been added.
- [x] Documentation has been added.