Use poetry 2.1.2 in the devcontainer (#5205)
### Motivation

When trying to update the `poetry.lock` in #5204 it uses poetry 1.6.1,
which leads to massive diff and may or may not work properly.

To deliver that PR I manually installed the right version, but it's not
the right way to go.

### Implementation

Update the devconainer config to use poetry 2.1.2, as in the
`poetry.lock`


https://github.com/zauberzeug/nicegui/blob/987f59ff88076ad4a9c6078618044b791e6cd4a5/poetry.lock#L1

(Also the git blame is fun. The version oscilates between Dependabot and
human commits)

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.