Fix support for Redis Unix-Domainsocket-connections (#5214)
### Motivation

I usually use Redis with UnixDomainsocket-connections, using
connection-URLs beginning with `unix://`. When trying to use this for
persistence in NiceGUI I discovered, that this throws the error 'Could
not load data from Redis with key ...'. The reason is, that the option
`socket_keepalive` is not present when `redis.from_url()` leads to the
use of Redis' class `UnixDomainSocketConnection`.

### Implementation

The fix builds a connection-parameter-dict in the class
`RedisPersistentDict` and adds the option 'socket_keepalive' ony when
the URL is not beginning with `unix://`. Furthermore it reuses this dict
for the Redis-connection in the same Class on `initialize_sync()`.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are are not necessary.
- [x] Documentation is not necessary.

---------

Co-authored-by: Rodja Trappe <rodja@zauberzeug.com>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>