Fix emitting and calling chained events (#5238)
### Motivation

While migrating [RoSys](https://github.com/zauberzeug/rosys/) to NiceGUI
3.0, we noticed that event chaining doesn't work if the handler expects
arguments:

```py
event1 = Event[str]()
event2 = Event[str]()
ui.button('Click me', on_click=lambda: event1.emit('Hello'))
event1.subscribe(event2.emit)
event2.subscribe(ui.notify)
```

### Implementation

When emitting an event, the function `helpers.expects_arguments` is used
to check if the handler needs to be called with an argument. But the
`emit` function itself expects arbitrarily many arguments, so it is
unclear how to call it. The caller of `emit` can't know when the
internally called subscribers expect.

Therefore we added an extra check if the callback is an `emit` or `call`
method. In this case, arguments are provided.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation is not necessary.